Added parameters list in dmisc + Functions displacement
More practical to have access to the list using hub.dmisc['parameters']

Also moved toolbox functions from _init_ to a dedicated toolbox file
